+++ docs/changes.rst
@@ -24,6 +24,8 @@ Unreleased.
      - Added :attr:`Image.colors <wand.image.BaseImage.colors>` property.
      - Added :meth:`Image.color_map() <wand.image.BaseImage.color_map>` method.
      - Added :meth:`Image.cycle_color_map() <wand.image.BaseImage.cycle_color_map>` method.
+ - Support for ``highlight`` & ``lowlight`` has been added to
+   :meth:`Image.compare() <wand.image.BaseImage.compare>` method.
  - Added :meth:`Image.black_threshold() <wand.image.BaseImage.black_threshold>` method.
  - Added :meth:`Image.blue_shift() <wand.image.BaseImage.blue_shift>` method.
  - Added :meth:`Image.color_matrix() <wand.image.BaseImage.color_matrix>` method.